Create tab_untab_vim.json
open keyboard shortcuts (Ctrl+K Ctrl+S) from settings. At the end of navigation bar or the section where new tabs appears there will be symbol, a page with an arrow on it top left corner click on the symbol and a json file will appear then paste the code.